+++ script-utils-java/src/main/kotlin/com/pambrose/common/script/JavaScriptPool.kt
@@ -16,13 +16,13 @@
 
 package com.pambrose.common.script
 
-import kotlinx.coroutines.runBlocking
-
 /**
  * A pre-populated pool of [JavaScript] instances backed by a coroutine [Channel][kotlinx.coroutines.channels.Channel].
  *
  * Instances are created eagerly during initialization and recycled with context resets
  * after each use.
+ *
+ * Closing the pool closes its instances.
  *
  * @param size the number of [JavaScript] instances to create in the pool
  * @param nullGlobalContext ignored: [JavaScript] binds variables into the engine scope, never the
@@ -34,8 +34,6 @@ class JavaScriptPool(
   @Suppress("UNUSED_PARAMETER", "UnusedPrivateProperty") nullGlobalContext: Boolean = false,
 ) : AbstractScriptPool<JavaScript>(size, false) {
   init {
-    runBlocking {
-      repeat(size) { channel.send(JavaScript()) }
-    }
+    populate { JavaScript() }
   }
 }