+++ script-utils-kotlin/src/main/kotlin/com/pambrose/common/script/KotlinExprEvaluatorPool.kt
@@ -16,13 +16,13 @@
 
 package com.pambrose.common.script
 
-import kotlinx.coroutines.runBlocking
-
 /**
  * A pre-populated pool of [KotlinExprEvaluator] instances backed by a coroutine
  * [Channel][kotlinx.coroutines.channels.Channel].
  *
  * Instances are created eagerly during initialization.
+ *
+ * Closing the pool closes its instances.
  *
  * @param size the number of [KotlinExprEvaluator] instances to create in the pool
  */
@@ -30,8 +30,6 @@ class KotlinExprEvaluatorPool(
   size: Int,
 ) : AbstractExprEvaluatorPool<KotlinExprEvaluator>(size) {
   init {
-    runBlocking {
-      repeat(size) { channel.send(KotlinExprEvaluator()) }
-    }
+    populate { KotlinExprEvaluator() }
   }
 }