+++ script-utils-python/src/main/kotlin/com/pambrose/common/script/PythonExprEvaluatorPool.kt
@@ -16,13 +16,13 @@
 
 package com.pambrose.common.script
 
-import kotlinx.coroutines.runBlocking
-
 /**
  * A pre-populated pool of [PythonExprEvaluator] instances backed by a coroutine
  * [Channel][kotlinx.coroutines.channels.Channel].
  *
  * Instances are created eagerly during initialization.
+ *
+ * Closing the pool closes its instances.
  *
  * @param size the number of [PythonExprEvaluator] instances to create in the pool
  */
@@ -30,8 +30,6 @@ class PythonExprEvaluatorPool(
   size: Int,
 ) : AbstractExprEvaluatorPool<PythonExprEvaluator>(size) {
   init {
-    runBlocking {
-      repeat(size) { channel.send(PythonExprEvaluator()) }
-    }
+    populate { PythonExprEvaluator() }
   }
 }