+++ script-utils-python/src/main/kotlin/com/pambrose/common/script/PythonScriptPool.kt
@@ -16,14 +16,14 @@
 
 package com.pambrose.common.script
 
-import kotlinx.coroutines.runBlocking
-
 /**
  * A pre-populated pool of [PythonScript] instances backed by a coroutine
  * [Channel][kotlinx.coroutines.channels.Channel].
  *
  * Instances are created eagerly during initialization and recycled with context resets
  * after each use.
+ *
+ * Closing the pool closes its instances.
  *
  * @param size the number of [PythonScript] instances to create in the pool
  * @param nullGlobalContext if `true`, resets the global scope bindings to `null` when recycling
@@ -33,8 +33,6 @@ class PythonScriptPool(
   nullGlobalContext: Boolean,
 ) : AbstractScriptPool<PythonScript>(size, nullGlobalContext) {
   init {
-    runBlocking {
-      repeat(size) { channel.send(PythonScript(nullGlobalContext)) }
-    }
+    populate { PythonScript(nullGlobalContext) }
   }
 }